+++ pkg/util/k8sutil/pod_util.go
@@ -127,6 +127,9 @@ func applyPodPolicy(clusterName string, pod *v1.Pod, policy *spec.PodPolicy) {
 	if len(policy.Tolerations) != 0 {
 		pod.Spec.Tolerations = policy.Tolerations
 	}
+	if policy.AutomountServiceAccountToken != nil {
+		pod.Spec.AutomountServiceAccountToken = policy.AutomountServiceAccountToken
+	}
 
 	mergeLabels(pod.Labels, policy.Labels)
 
@@ -150,6 +153,9 @@ func applyPodPolicyToPodTemplateSpec(clusterName string, pod *v1.PodTemplateSpec
 	}
 	if len(policy.Tolerations) != 0 {
 		pod.Spec.Tolerations = policy.Tolerations
+	}
+	if policy.AutomountServiceAccountToken != nil {
+		pod.Spec.AutomountServiceAccountToken = policy.AutomountServiceAccountToken
 	}
 
 	mergeLabels(pod.Labels, policy.Labels)