+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Expose `/metrics` endpoint at port 8080
 - Add cluster S3 spec `prefix` feature. Let user choose a prefix under the bucket.
+- Add `automountServiceAccountToken` to pod policy. Let users disable automounting of the Kubernetes access token into etcd-operator controlled pods.
 - Cluster backups can now be saved using Azure Blob Storage (ABS).
 
 ### Changed